@@ -1,23 +1,30 @@
-import logo from './logo.svg';
-import './App.css';
+import "./App.css";
+import { useQuery } from "react-query";
+import { useTransition } from "react";
 
 function App() {
+  const [isPending, startTransition] = useTransition();
+  const result = useQuery(
+    "foo",
+    async () => {
+      await new Promise((res) => setTimeout(res, 1000));
+      return () => "foo" + Math.random();
+    },
+    { suspense: true }
+  );
+  console.log(result);
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      {JSON.stringify(result)}
+      <button
+        onClick={() => {
+          startTransition(() => {
+            result.refetch();
+          });
+        }}
+      >
+        fetch
+      </button>
     </div>
   );
 }
